requirements.txt
streamlit
pandas
google-genai
numpy
matplotlib
seaborn
openpyxl